requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
azure-core==1.35.1
azure-storage-blob==12.26.0
babel==2.17.0
bcrypt==4.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.2
fastapi==0.116.1
fastapi-admin==1.0.4
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
isodate==0.7.2
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pendulum==3.1.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
pypika-tortoise==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
